proyectos_python/generar_segmentos/requirements.txt
snowflake-connector-python==3.0.4